+++ skia/src/gen/java/org/bytedeco/skia/sk_region_spanerator_t.java
@@ -10,11 +10,10 @@ import static org.bytedeco.javacpp.presets.javacpp.*;
 
 import static org.bytedeco.skia.global.Skia.*;
 
-
 @Opaque @Properties(inherit = org.bytedeco.skia.presets.Skia.class)
-public class gr_glinterface_t extends Pointer {
+public class sk_region_spanerator_t extends Pointer {
     /** Empty constructor. Calls {@code super((Pointer)null)}. */
-    public gr_glinterface_t() { super((Pointer)null); }
+    public sk_region_spanerator_t() { super((Pointer)null); }
     /** Pointer cast constructor. Invokes {@link Pointer#Pointer(Pointer)}. */
-    public gr_glinterface_t(Pointer p) { super(p); }
+    public sk_region_spanerator_t(Pointer p) { super(p); }
 }